requirements.txt
apache-airflow==2.8.1
requests==2.31.0
kafka-python==2.0.2